refactor: movies에서 home으로 네이밍 변경
