+++ app/src/main/java/woowacourse/movie/feature/home/presenter/HomePresenter.kt
@@ -1,15 +1,15 @@
-package woowacourse.movie.feature.movies.presenter
+package woowacourse.movie.feature.home.presenter
 
 import woowacourse.movie.domain.model.Movie.Companion.movies
 import woowacourse.movie.domain.model.Screening
 import woowacourse.movie.domain.model.Screenings.Companion.screenings
+import woowacourse.movie.feature.home.contract.HomeContract
 import woowacourse.movie.feature.mapper.toUi
 import woowacourse.movie.feature.model.MovieUiModel
-import woowacourse.movie.feature.movies.contract.MoviesContract
 
-class MoviesPresenter(
-    private val view: MoviesContract.View,
-) : MoviesContract.Presenter {
+class HomePresenter(
+    private val view: HomeContract.View,
+) : HomeContract.Presenter {
     override fun prepareMovies() {
         view.showMovies(movies.map { it.toUi() })
     }
