+++ app/src/main/java/woowacourse/movie/feature/home/view/adapter/Item.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.feature.movies.view.adapter
+package woowacourse.movie.feature.home.view.adapter
 
 import woowacourse.movie.feature.model.MovieUiModel
