+++ app/src/main/java/woowacourse/movie/feature/home/view/adapter/MoviesAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.feature.movies.view.adapter
+package woowacourse.movie.feature.home.view.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
