+++ app/src/main/res/layout/fragment_booking_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.movies.view.HomeFragment">
+    tools:context=".feature.home.view.HomeFragment">
 
     <TextView
         android:layout_width="match_parent"
